refactor: cleanup types module
