+++ src/types/params/plain_secret.rs
@@ -7,6 +7,7 @@ use ::rsa::traits::PrivateKeyParts;
 use byteorder::{BigEndian, ByteOrder};
 use bytes::{BufMut, BytesMut};
 use hkdf::Hkdf;
+use log::debug;
 use num_bigint::ModInverse;
 use sha2::Sha256;
 use zeroize::ZeroizeOnDrop;
